Refactor existing form submit buttons to make them easier to reuse
