--- support-frontend/assets/components/subscriptionCheckouts/subscriptionSubmitButtons.jsx
@@ -1,100 +1,0 @@
-// @flow
-
-// ----- Imports ----- //
-
-import type { Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
-import React from 'react';
-
-import { type IsoCurrency } from 'helpers/internationalisation/currency';
-import { hiddenIf } from 'helpers/utilities';
-import { type SetupPayPalRequestType } from 'helpers/paymentIntegrations/payPalRecurringCheckout';
-import type { BillingPeriod } from 'helpers/billingPeriods';
-import { PayPalExpressButton } from 'components/paypalExpressButton/PayPalExpressButton';
-import Button from 'components/button/button';
-import { type Option } from 'helpers/types/option';
-import type { PaymentMethod } from 'helpers/paymentMethods';
-import { PayPal } from 'helpers/paymentMethods';
-import { type FormError } from 'helpers/subscriptionsForms/validation';
-import { type FormField } from 'helpers/subscriptionsForms/formFields';
-import Heading from 'components/heading/heading';
-
-import 'components/forms/customFields/error.scss';
-import './subscriptionSubmitButton.scss';
-
-// ----- Types ----- //
-
-type PropTypes = {|
-  paymentMethod: Option<PaymentMethod>,
-  currencyId: IsoCurrency,
-  csrf: CsrfState,
-  setupRecurringPayPalPayment: SetupPayPalRequestType,
-  payPalHasLoaded: boolean,
-  isTestUser: boolean,
-  onPaymentAuthorised: Function,
-  amount: number,
-  billingPeriod: BillingPeriod,
-  validateForm: Function,
-  formIsValid: Function,
-  allErrors: FormError<FormField>[],
-|};
-
-type ErrorSummaryPropTypes = {
-  errors: Array<Object>,
-}
-
-const ErrorSummary = (props: ErrorSummaryPropTypes) => (
-  <div className="component-form-error__border">
-    <Heading className="component-form-error__heading" size={2}>There is a problem</Heading>
-    <ul>
-      {props.errors.map(error => (
-        <li className="component-form-error__summary-error">
-          {error.message}
-        </li>
-      ))}
-    </ul>
-  </div>
-
-);
-
-// ----- Render ----- //
-
-
-function SubscriptionSubmitButtons(props: PropTypes) {
-  // We have to show/hide PayPalExpressButton rather than conditionally rendering it
-  // because we don't want to destroy and replace the iframe each time.
-  // See PayPalExpressButton for more info.
-  return (
-    <div className="component-submit-button">
-      <div className="component-submit-button--margin">
-        <div
-          id="component-paypal-button-checkout"
-          className={hiddenIf(props.paymentMethod !== PayPal, 'component-paypal-button-checkout')}
-        >
-          <PayPalExpressButton
-            onPaymentAuthorisation={props.onPaymentAuthorised}
-            csrf={props.csrf}
-            currencyId={props.currencyId}
-            hasLoaded={props.payPalHasLoaded}
-            canOpen={props.formIsValid}
-            onClick={props.validateForm}
-            formClassName="form--contribution"
-            isTestUser={props.isTestUser}
-            setupRecurringPayPalPayment={props.setupRecurringPayPalPayment}
-            amount={props.amount}
-            billingPeriod={props.billingPeriod}
-          />
-        </div>
-        <Button
-          id="qa-submit-button"
-          type="submit"
-          modifierClasses={props.paymentMethod === PayPal ? ['hidden'] : []}
-        >
-          Continue to payment
-        </Button>
-      </div>
-      <span>{props.allErrors.length > 0 && <ErrorSummary errors={props.allErrors} />}</span>
-    </div>
-  );
-}
-
-export { SubscriptionSubmitButtons };
